Scrub through chart

+++ app/src/main/java/com/veyndan/bitcoin/data/BitcoinChart.kt
@@ -1,5 +1,7 @@
 package com.veyndan.bitcoin.data
 
+import java.math.BigDecimal
+
 data class BitcoinChart(
     val name: String,
     val values: List<Datapoint>
@@ -7,5 +9,5 @@ data class BitcoinChart(
 
 data class Datapoint(
     val x: Long,
-    val y: Double
+    val y: BigDecimal
 )
